ai-sre/requirements.txt
crewai
litellm
groq
mistralai
python-dotenv
kubernetes
pydantic
prometheus-client